requirements.txt
flask
flask-login
flask-limiter
flask-wtf
python-dotenv
werkzeug
pillow